Export strip_xscope in utils
